Move entrypoint macro impl to the separate module
